+++ core/src/test/java/com/backblaze/b2/client/structures/B2EventNotificationRuleTest.java
@@ -40,6 +40,7 @@ public class B2EventNotificationRuleTest extends B2BaseTest {
                 "      }\n" +
                 "    ],\n" +
                 "    \"hmacSha256SigningSecret\": \"3XDfkdQte2OgA78qCtSD17LAzpj6ay9H\",\n" +
+                "    \"maxEventsPerBatch\": 20,\n" +
                 "    \"targetType\": \"webhook\",\n" +
                 "    \"url\": \"https://www.example.com\"\n" +
                 "  }\n" +
@@ -66,7 +67,8 @@ public class B2EventNotificationRuleTest extends B2BaseTest {
                                                 new B2WebhookCustomHeader("name2", "val2")
                                         )
                                 ),
-                                "3XDfkdQte2OgA78qCtSD17LAzpj6ay9H"
+                                "3XDfkdQte2OgA78qCtSD17LAzpj6ay9H",
+                                20
                         ),
                         true,
                         false,